+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change log
 
+## [Unreleased]
+
+### Changed
+
+- concord-server: fixed retrieval of team users.
+
+
+
 ## [0.57.0] - 2017-02-22
 
 ### Added